requirements.txt
django
whitenoise
langchain
langchain-community
PyPDF2
pytesseract
tesseract
python-dotenv
django-vectordb
pdf2image
google-generativeai
langchain-google-genai
openai
pyyaml
sentence-transformers
markdown
gunicorn
azure-storage-blob
django-storages
ghostscript
tabula
tabula-py
JPype1
pandas